ZC moment-order choice now limitied via dropdown menu
The corresponding chi coefficients to moment-order (1:20,25,30) are shipped with the app, but have to be computed if the user choose a different moment-order - this can take too long time and will create a bad user experinece. This limitation can be fixed in the future, but it will be good enough for most use-cases.

Former-commit-id: b176f2b82fd30de40eb864f66508c63d7062a9dc